% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add.buffer}
\alias{add.buffer}
\title{Add buffer}
\usage{
add.buffer(grid)
}
\arguments{
\item{grid}{Grid of class \code{momo.grid} as returned by the function
\link{create.grid}.}
}
\value{
A list of class \code{momo.grid} with a information about the spatial
domain and grid.
}
\description{
Add a buffer zone around a grid of class \code{momo.grid}.
}
