% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{add.report}
\alias{add.report}
\title{Add TMB's report}
\usage{
add.report(fit)
}
\arguments{
\item{fit}{A fitted momo list of class \code{momo.fit} as returned by the
function \link{fit.momo}.}
}
\value{
A list of class \code{momo.fit} as returned by the function \link{fit.momo}
with reported quantites.
}
\description{
Run and add TMB's reporting function and add it to the fitted
list.
}
