% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{add.sdreport}
\alias{add.sdreport}
\title{Add TMB's sdreport}
\usage{
add.sdreport(fit)
}
\arguments{
\item{fit}{A fitted momo list of class \code{momo.fit} as returned by the
function \link{fit.momo}.}
}
\value{
A list of class \code{momo.fit} as returned by the function \link{fit.momo}
with sdreport.
}
\description{
Run and add TMB's sdreporting function and add it to the fitted
list.
}
