% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BankHolidays.R
\name{BankHolidays}
\alias{BankHolidays}
\title{Bank holidays or intensity of registration of deaths
Substitude the demanded BankHoliday file}
\usage{
BankHolidays(MData)
}
\arguments{
\item{MData}{A data.frame or data.table with MOMO input data}
}
\value{
Data with registration intensity (between 0 and 1) per date
}
\description{
Bank holidays or intensity of registration of deaths
Substitude the demanded BankHoliday file
}
