% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmark.R
\name{BenchmarkDelay}
\alias{BenchmarkDelay}
\title{Benchmark different delay corrections.}
\usage{
BenchmarkDelay(tag, wdir = NULL)
}
\arguments{
\item{tag}{Tag for file name}

\item{wdir}{workingdir}
}
\description{
Benchmark different delay corrections.
}
