% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create.grid}
\alias{create.grid}
\title{Create a grid}
\usage{
create.grid(
  xrange = c(0, 1),
  yrange = c(0, 1),
  dxdy = c(0.1, 0.1),
  select = FALSE,
  plot.land = FALSE,
  keep.gpar = FALSE,
  verbose = TRUE,
  fit = NULL,
  grid = NULL
)
}
\arguments{
\item{xrange}{range of the x-dimension of spatial domain. Default: \code{c(0,1)}.}

\item{yrange}{range of the y-dimension of spatial domain. Default: \code{c(0,1)}.}

\item{dxdy}{resolution of grid in x and y direction. Default: \code{c(0.1,0.1)}.}

\item{select}{logical; if \code{TRUE}, allows to select cells in spatial grid.
Default: \code{FALSE}. If a vector of length > 1 is provided, the vector is
interpreted as an index of the selected grid cells of the full grid.}

\item{plot.land}{plot.land logical; If \code{TRUE}, plot land masses using the
function \link[maps:map]{maps::map}. Default: \code{FALSE}.}

\item{keep.gpar}{logical; If \code{TRUE}, do not overwrite the graphical
parameters. Default: \code{FALSE}.}

\item{verbose}{if \code{TRUE}, print information to console. Default: \code{TRUE}.}

\item{fit}{optional; allows to extract and update grid from fitted object of
class \code{momo.fit} returned by function \link{fit.momo}. Default: \code{NULL}. NULL.}

\item{grid}{optional; allows to update a grid of class \code{momo.grid} as
returned by the function \link{create.grid}. Default: \code{NULL}.}
}
\value{
A list of class \code{momo.grid} with a information about the spatial
domain and grid.
}
\description{
\code{create.grid} allows to create a new grid or modify an existing
grid that is used by \link{fit.momo} to predict movement rates and required
for the matrix exponential approach.
}
\examples{
grid <- create.grid()

}
