% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_MOMOmaster_defaults.R
\docType{data}
\name{dataExport}
\alias{dataExport}
\title{Data to be exported}
\format{An object of class \code{environment} of length 0.}
\usage{
dataExport
}
\description{
Data to be exported
}
\keyword{datasets}
