% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{def.conf}
\alias{def.conf}
\title{def.conf}
\usage{
def.conf(dat)
}
\arguments{
\item{dat}{Dat frame with required information created by function \code{\link{check.momo.data}}}

\item{const.dif}{Constant diffusion? logical}
}
\description{
def.conf
}
