% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{def.map}
\alias{def.map}
\title{def.map}
\usage{
def.map(dat, conf, par)
}
\arguments{
\item{dat}{data frame with input data as produced by the function
\link{check.momo.data}.}

\item{conf}{configuration list as produced by the function \link{def.conf}.}

\item{par}{parameter list with initial values as produced by the function
\link{def.par}.}
}
\description{
Get a list with default mappings for parameters.
}
\examples{
data(skjepo)

map <- with(skjepo, def.map(dat, conf, par))

}
