% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{def.par}
\alias{def.par}
\title{def.par}
\usage{
def.par(dat, conf)
}
\arguments{
\item{dat}{data frame with input data as produced by the function
\link{check.momo.data}.}

\item{conf}{configuration list as produced by the function \link{def.conf}.}
}
\value{
A named list.
}
\description{
Get a list with parameters and default initial values.
}
\examples{
data(skjepo)

par <- with(skjepo, def.par(dat, conf))

}
