% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_original.R
\name{delayMOMO_original}
\alias{delayMOMO_original}
\title{Delay adjustment
as originally coded.}
\usage{
delayMOMO_original(aggr, zvalue = 1.96)
}
\arguments{
\item{aggr}{Output data from aggregateMOMO()}

\item{zvalue}{95\% z-value}
}
\description{
Delay adjustment
as originally coded.
}
