% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_richard.R
\name{delayMOMO_richard}
\alias{delayMOMO_richard}
\title{delayMOMO_richard
This is the package standardly used for Norway
It has three parts.
For delay=N:2, it uses the original method.
For delay=1, it uses the prediction from delay=2 and WR1
For delay=0, it uses the prediction from delay=2 and WR0}
\usage{
delayMOMO_richard(aggr, zvalue = 1.96)
}
\arguments{
\item{aggr}{Output data from aggregateMOMO()}

\item{zvalue}{95\% z-value}
}
\description{
delayMOMO_richard
This is the package standardly used for Norway
It has three parts.
For delay=N:2, it uses the original method.
For delay=1, it uses the prediction from delay=2 and WR1
For delay=0, it uses the prediction from delay=2 and WR0
}
