% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_sincos.R
\name{delayMOMO_sincos}
\alias{delayMOMO_sincos}
\title{Delay adjustment as originally coded
but with a sin curve over time.}
\usage{
delayMOMO_sincos(aggr, zvalue = 1.96)
}
\arguments{
\item{aggr}{Output data from aggregateMOMO()}

\item{zvalue}{95\% z-value}
}
\description{
Delay adjustment as originally coded
but with a sin curve over time.
}
