% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{do.sdreport}
\alias{do.sdreport}
\title{Do sd report}
\usage{
do.sdreport(fit, ignore.parm.uncertainty = FALSE)
}
\arguments{
\item{fit}{fit}
}
\value{
list of result of minimization conducted via nlminb
}
\description{
Do sd report
}
\details{
...
}
\section{@method sdreport momo.fit}{
\subsection{TODO: why does the method sdreport.momo.fit not work?}{
}
}

\examples{
3 + 3
}
