% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.momo}
\alias{fit.momo}
\title{Fit the movement model}
\usage{
fit.momo(
  dat,
  conf = NULL,
  par = NULL,
  map = NULL,
  run = TRUE,
  lower = get.lower.bounds(par),
  upper = get.upper.bounds(par),
  rel.tol = 1e-10,
  do.sdreport = TRUE,
  do.report = TRUE,
  use.expm = NULL,
  use.rel.events = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{data frame with input data as produced by the function
\link{check.momo.data}.}

\item{conf}{configuration list as produced by the function \link{def.conf}. If
\code{NULL} (default), \link{def.conf} is used to generate the default
configuration list based on the input data.}

\item{par}{parameter list with initial values as produced by the function
\link{def.par}. If \code{NULL} (default), \link{def.par} is used to generate the
default parameter list based on the input data and configuration list.}

\item{map}{list with parameter mapping as produced by the function
\link{def.map}. If \code{NULL} (default), \link{def.map} is used to generate the
default mapping list based on the input data and configuration and
parameter lists.}

\item{run}{logical; If \code{FALSE}, the AD object is returned with the
optimization. Default: \code{FALSE}.}

\item{lower}{by default, \link{get.lower.bounds} is used.}

\item{upper}{by default, \link{get.upper.bounds} is used.}

\item{rel.tol}{option passed to \link[stats:nlminb]{stats::nlminb} sets the convergence
criteria. Default: \code{1e-10}.}

\item{do.sdreport}{logical; If \code{FALSE}, \link[RTMB:TMB-interface]{RTMB::sdreport} is not run and no
parameter uncertainties are returned. Default: \code{TRUE}.}

\item{do.report}{logical; If \code{FALSE}, \code{obj$report()} is not run and RTMB
variables might not be reported. Default: \code{TRUE}.}

\item{use.expm}{logical; allows to overwrite setting in configuration list
(\code{conf$use.expm}). If \code{NULL}, setting from configuration list is used.
Default: \code{NULL}.}

\item{use.rel.events}{logical; allows to overwrite setting in configuration
list (\code{conf$use.rel.events}). If \code{NULL}, setting from configuration list
is used. Default: \code{NULL}.}

\item{verbose}{if \code{TRUE}, print information to console. Default: \code{TRUE}.}

\item{...}{extra arguments to \link[RTMB:TMB-interface]{RTMB::MakeADFun}.}
}
\value{
A list of class \code{momo.fit}.
}
\description{
The movement model, momo, allows estimation of habitat
preferences and fine-scale movement rates using various types of tagging
data. It supports flexible spatiotemporal modeling with options for
Kalman filter or matrix exponential approaches, and includes tools for
data preparation, model fitting, and visualization.
}
\examples{
data(skjepo)

fit <- fit.momo(skjepo)

}
