% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.cfl}
\alias{get.cfl}
\title{Get Courant–Friedrichs–Lewy condition}
\usage{
get.cfl(grid, env, par)
}
\arguments{
\item{grid}{a grid object of class \code{momo.grid} as returned by the function
\link{create.grid}.}

\item{env}{a list with environmental covariates of class \code{momo.env} as
returned by the function \link{prep.env}.}

\item{par}{parameter list with initial values as produced by the function
\link{def.par}.}
}
\value{
The Courant-Friedrichs-Lewy condition.
}
\description{
\code{get.cfl} allows to calculate the Courant-Friedrichs-Lewy
condition.
}
