% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.neighbours}
\alias{get.neighbours}
\title{Get neighbours for each cell of a grid}
\usage{
get.neighbours(celltable, diagonal = FALSE)
}
\arguments{
\item{celltable}{celltable}
}
\description{
Get neighbours for each cell of a grid
}
