% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.precision.matrix}
\alias{get.precision.matrix}
\title{Get precision matrix}
\usage{
get.precision.matrix(grid, h, nu, rho, delta, matern = TRUE, diagonal = FALSE)
}
\arguments{
\item{grid}{a grid object of class \code{momo.grid} as returned by the function
\link{create.grid}.}

\item{h}{parameter of the matern covariance structure.}

\item{nu}{parameter of the matern covariance structure.}

\item{rho}{parameter of the matern covariance structure.}

\item{delta}{parameter of the matern covariance structure.}

\item{matern}{logical; if \code{TRUE} (default), matern covariance structure is
used for simulation of environmental fields.}

\item{diagonal}{logical; if \code{TRUE}, diagonal neighbours are considered in
neighbouring structure. Default: \code{FALSE}.}
}
\value{
A precision matrix.
}
\description{
\code{get.precision.matrix} allows to create a precision matrix
based on a grid.
}
