% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.release.events}
\alias{get.release.events}
\title{Get release events}
\usage{
get.release.events(dat, grid = NULL, time.cont = NULL, age.max = NULL)
}
\arguments{
\item{dat}{}
}
\description{
Get release events
}
