% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{get.upper.bounds}
\alias{get.upper.bounds}
\title{Upper bounds}
\usage{
get.upper.bounds(par)
}
\arguments{
\item{par}{parameter list with initial values as produced by the function
\link{def.par}.}
}
\value{
List with upper bounds for parameters.
}
\description{
Get upper bounds for parameters.
}
