% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habi.R
\name{habi.full}
\alias{habi.full}
\title{Habitat function}
\usage{
habi.full(FIELDS, XR, YR, ienv, time.cont, S, dS, ienvS)
}
\arguments{
\item{FIELDS}{fields}
}
\value{
Different things
}
\description{
Habitat function
}
\details{
habitat function
}
