% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momo.R
\name{JN_plot}
\alias{JN_plot}
\title{Generate Johnson-Neyman Figure from momo object}
\usage{
JN_plot(
  momoObject,
  ci = "bc",
  min_z = -3,
  max_z = 3,
  detail = 300,
  lower.quantile = 0.025,
  upper.quantile = 0.975,
  alpha = 0.2,
  sd.line = 2
)
}
\arguments{
\item{momoObject}{momo object (output from momo_modsem or momo_mplus).}

\item{min_z}{minimum values of moderator Z on the graph.}

\item{max_z}{maximum values of moderator Z on the graph.}

\item{detail}{detail of the Johnson-Neyman graph. The higher the number, the smoother the lines, but it will also take longer to generate.}

\item{lower.quantile}{lower quantile for the confidence intervals. 0.025 for 95-percent confidence intervals (default). 0.005 for 99-percent confidence intervals.}

\item{upper.quantile}{upper quantile for the confidence intervals. 0.975 for 95-percent confidence intervals (default). 0.995 for 99-percent confidence intervals.}

\item{alpha}{alpha = 0.2 (default) Thickness of the lines.}

\item{sd.line}{horiznotal line that shows mean +/- 2 (default) values of the moderator.}
}
\value{
a figure for changing the titles.
}
\description{
Generate Johnson-Neyman Figure from momo object
}
\examples{

## -- Example -- ##

# momoObject is mcObject & output JN_plot object is JN_figure

JN_figure <- JN_plot(mcObject)

}
