% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{model.version.info}
\alias{model.version.info}
\title{Description of model}
\usage{
model.version.info(fit, ...)
}
\arguments{
\item{fit}{returned object from fit.movemod}

\item{...}{Additional parameters to be passed to ...}
}
\description{
Description of model
}
\details{
Writes a string to install the version of the package which was
used to run the model.
}
