% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{momo.cols}
\alias{momo.cols}
\title{Momo colors}
\usage{
momo.cols(n = 1, alpha = 1, type = NULL)
}
\arguments{
\item{n}{number of colors. Default: \code{1}.}

\item{alpha}{transparency value. Default: \code{1}.}

\item{type}{optional; allows to extract colors for specific occasions.
Default: \code{NULL}.}
}
\value{
A vector with colors.
}
\description{
\code{momo.cols} returns a vector with colors for different
scenarios.
}
