% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momo.R
\name{momo_lavaan_fun}
\alias{momo_lavaan_fun}
\title{Monte Carlo Simulation for Confidence Intervals of Defined Function (lavaan)}
\usage{
momo_lavaan_fun(object, Sfunction = "NULL", R = 5)
}
\arguments{
\item{object}{lavaan object (output from lavaan).}

\item{Sfunction}{function of estimated parameters from lavaan object.}

\item{R}{number of Monte Carlo simulation samples (in millions). For example, R=5 (default) generates 5,000,000 simulated samples.}
}
\value{
momo confidence intervals.
}
\description{
Generate confidence intervals of defined function from lavaan results using Monte Carlo simulation.
}
\examples{

## -- Example -- ##

# lavaan object is "est_lms"

momo_lavaan_fun(est_lms, Sfunction = "a1*a2")

}
