% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momo.R
\name{momo_modsem_fun}
\alias{momo_modsem_fun}
\title{Monte Carlo Simulation for Confidence Intervals of Defined Function (modsem_da)}
\usage{
momo_modsem_fun(object, Sfunction = "NULL", R = 5)
}
\arguments{
\item{object}{modsem object (output from modsem).}

\item{Sfunction}{function of estimated parameters from modsem object.}

\item{R}{number of Monte Carlo simulation samples (in millions). For example, R=5 (default) generates 5,000,000 simulated samples.}
}
\value{
momo confidence intervals.
}
\description{
Generate confidence intervals of defined function from modsem da approaches (lms or qml) results using Monte Carlo simulation.
}
\examples{

## -- Example -- ##

# modsem object is "est_lms"

momo_modsem_fun(est_lms, Sfunction = "a1*a2")

}
