% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momo.R
\name{momo_mplus_fun}
\alias{momo_mplus_fun}
\title{Monte Carlo Simulation for Confidence Intervals of Defined Function (mplus)}
\usage{
momo_mplus_fun(
  mplus_output_file = "mplus_output.out",
  a1 = "NA",
  a2 = "NA",
  a3 = "NA",
  a4 = "NA",
  z1 = "NA",
  z2 = "NA",
  z3 = "NA",
  z4 = "NA",
  w1 = "NA",
  w2 = "NA",
  w3 = "NA",
  w4 = "NA",
  zw1 = "NA",
  zw2 = "NA",
  zw3 = "NA",
  zw4 = "NA",
  Sfunction = "NULL",
  R = 5
)
}
\arguments{
\item{mplus_output_file}{Mplus output (.out) file (output from Mplus).}

\item{a1}{parameter name of a1 in MODEL CONSTRAINT: NEW.}

\item{a2}{parameter name of a2 in MODEL CONSTRAINT: NEW.}

\item{a3}{parameter name of a3 in MODEL CONSTRAINT: NEW.}

\item{a4}{parameter name of a4 in MODEL CONSTRAINT: NEW.}

\item{z1}{parameter name of z1 in MODEL CONSTRAINT: NEW.}

\item{z2}{parameter name of z2 in MODEL CONSTRAINT: NEW.}

\item{z3}{parameter name of z3 in MODEL CONSTRAINT: NEW.}

\item{z4}{parameter name of z4 in MODEL CONSTRAINT: NEW.}

\item{w1}{parameter name of w1 in MODEL CONSTRAINT: NEW.}

\item{w2}{parameter name of w2 in MODEL CONSTRAINT: NEW.}

\item{w3}{parameter name of w3 in MODEL CONSTRAINT: NEW.}

\item{w4}{parameter name of w4 in MODEL CONSTRAINT: NEW.}

\item{zw1}{parameter name of zw1 in MODEL CONSTRAINT: NEW.}

\item{zw2}{parameter name of zw2 in MODEL CONSTRAINT: NEW.}

\item{zw3}{parameter name of zw3 in MODEL CONSTRAINT: NEW.}

\item{zw4}{parameter name of zw4 in MODEL CONSTRAINT: NEW.}

\item{Sfunction}{function of estimated parameters from Mplus outputs.}

\item{R}{number of Monte Carlo simulation samples (in millions). For example, R=5 (default) generates 5,000,000 simulated samples.}

\item{results_file}{Mplus a text file (.txt) that saves the Mplus results (RESULTS IS "filename.txt" in Mplus SAVEDATA:).}
}
\value{
confidence intervals of defined function
}
\description{
Generate confidence intervals of a defined function from Mplus results using Monte Carlo simulation.
Location of estimated parameters can be found in momo::TECH1().
}
\examples{

## -- Example -- ##

# mplus_output_file is "model cc4.out" 
# output momo object is mcObject

mcObject <- momo_mplus_fun("model cc4.out", a1 = "A1", a2 = "A2", Sfunction="a1*a2")

}
