% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_MOMOmaster_defaults.R
\docType{data}
\name{momoAttr}
\alias{momoAttr}
\title{Placeholder list for all the following options}
\format{An object of class \code{environment} of length 0.}
\usage{
momoAttr
}
\description{
Placeholder list for all the following options
}
\keyword{datasets}
