% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{nll}
\alias{nll}
\title{Negative log likelihood function}
\usage{
nll(par, dat)
}
\arguments{
\item{par}{List with initial values for the model in a format similar to
what is returned from the \link{def.par} function.}

\item{dat}{Data frame with input data as produced by the function
\link{check.momo.data}.}
}
\value{
The negative loglikelihood value.
}
\description{
Likelihood function for the movement model.
}
