% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.land}
\alias{plot.land}
\title{Plot land masses}
\usage{
\method{plot}{land}(
  xlim,
  ylim,
  shift = FALSE,
  col = adjustcolor(grey(0.7), 0.5),
  border = grey(0.5)
)
}
\arguments{
\item{xlim}{x limits}

\item{ylim}{y limits}

\item{shift}{logical; If \code{TRUE}, world map is shifted frmo \link{-180,180} to
\link{0,360} logitudinal range. Note, that polygons crossing the meridian are
excluded. Default: \code{FALSE}.}

\item{col}{Color of landmasses. Default: \code{adjustcolor(grey(0.7), 0.5)}.}

\item{border}{Color of borders. Default: \code{grey(0.5)}.}
}
\value{
Nothing.
}
\description{
Plot land masses.
}
