% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.atags}
\alias{plotmomo.atags}
\title{Plot momo atags}
\usage{
plotmomo.atags(
  atags,
  main = "Archival tags",
  keep.par = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\description{
Plot momo atags
}
