% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.compare.one}
\alias{plotmomo.compare.one}
\title{Plot to compare single quantity}
\usage{
plotmomo.compare.one(
  fit,
  ...,
  quantity = c("pref", "taxis", "dif", "par"),
  keep.par = FALSE,
  asp = 2,
  col = momo:::get.momo.cols(10),
  lty = 1:10,
  plot.legend = 1
)
}
\arguments{
\item{fit}{A result report as generated by running \code{\link{fit.momo}} or
a list containing objects fitted with spict.}

\item{...}{Optional additional momo fits.}

\item{quantity}{Quantities that can be compared. The following options are
currently implemented: \code{"taxis"} for the taxis, \code{"dif"} for the
diffusion, \code{"par"} for the parameters.}

\item{asp}{Positive number deining the target aspect ratio (columns / rows)
of the plot arrangement.}
}
\description{
Plot to compare single quantity
}
