% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.compare}
\alias{plotmomo.compare}
\title{Compare plot}
\usage{
plotmomo.compare(
  fit,
  ...,
  quantity = c("pref", "taxis", "dif", "par"),
  plot.land = FALSE,
  keep.gpar = FALSE,
  col = momo.cols(10),
  cor.dif = NULL,
  asp = 2,
  plot.legend = 1,
  bg = NULL
)
}
\arguments{
\item{fit}{a list of class \code{momo.fit} as returned by \link{fit.momo} or a list of
such (named) fitted objects.}

\item{...}{optional, additional momo fits.}

\item{quantity}{quantities that can be compared. The following options are
currently implemented: \code{"taxis"} for the taxis, \code{"dif"} for
the diffusion, \code{"par"} for the parameters.}

\item{plot.land}{logical; If \code{TRUE}, plot land masses using the function
\link[maps:map]{maps::map}. Default: \code{FALSE}.}

\item{keep.gpar}{logical; If \code{TRUE}, do not overwrite the graphical
parameters. Default: \code{FALSE}.}

\item{col}{color}

\item{asp}{positive number deining the target aspect ratio (columns / rows)
of the plot arrangement.}

\item{plot.legend}{logical or integer indicating which legend should be
plotted. Default: \code{1}.}

\item{bg}{background color of the graphical window. By default (\code{NULL}),
background is transparent.}

\item{cor}{scaling parameter for the arrows. By default (\code{NULL}), not used.}
}
\value{
Nothing.
}
\description{
Plotting function that allows to compare multiple fits or
fitted and simulated data.
}
