% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.ctags}
\alias{plotmomo.ctags}
\title{Plot momo ctags}
\usage{
plotmomo.ctags(
  ctags,
  main = "Conventional tags",
  plot.land = FALSE,
  keep.par = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\description{
Plot momo ctags
}
