% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.data}
\alias{plotmomo.data}
\title{Plot data}
\usage{
plotmomo.data(x, keep.gpar = FALSE, ...)
}
\arguments{
\item{x}{a list of class \code{momo.data} as returned by \link{setup.momo.data}.}

\item{keep.gpar}{logical; If \code{TRUE}, do not overwrite the graphical
parameters. Default: \code{FALSE}.}

\item{...}{additional arguments for the function \link{plot}.}
}
\value{
Nothing.
}
\description{
Plot data.
}
