% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.dif}
\alias{plotmomo.dif}
\title{Plot diffusion}
\usage{
plotmomo.dif(
  x,
  cor = 20,
  par = NULL,
  funcs = NULL,
  env = NULL,
  main = "Diffusion",
  alpha = 0.5,
  keep.par = FALSE,
  ...
)
}
\description{
Plot diffusion
}
