% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.env}
\alias{plotmomo.env}
\title{Plot momo env}
\usage{
plotmomo.env(
  env,
  main = "Environmental fields",
  labels = TRUE,
  keep.par = FALSE,
  plot.land = FALSE,
  xlab = "x",
  ylab = "y",
  ...
)
}
\description{
Plot momo env
}
