% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.pref.spatial}
\alias{plotmomo.pref.spatial}
\title{Plot spatial preference}
\usage{
plotmomo.pref.spatial(
  x,
  type = "taxis",
  col = hcl.colors(14, "YlOrRd", rev = TRUE),
  ci = 0.95,
  par = NULL,
  funcs = NULL,
  env = NULL,
  keep.par = FALSE,
  knots.tax = NULL,
  add = FALSE,
  ...
)
}
\description{
Plot spatial preference
}
