% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.taxis}
\alias{plotmomo.taxis}
\title{Plot taxis}
\usage{
plotmomo.taxis(
  x,
  select = NULL,
  average = TRUE,
  cor = 0.1,
  col = "black",
  alpha = 0.5,
  lwd = 1,
  main = "Taxis",
  xlab = "x",
  ylab = "y",
  xaxt = "s",
  yaxt = "s",
  plot.land = FALSE,
  keep.par = FALSE,
  add = FALSE,
  ...
)
}
\description{
Plot taxis
}
