% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.taxis2}
\alias{plotmomo.taxis2}
\title{Plot taxis2}
\usage{
plotmomo.taxis2(
  x,
  cor = 0.1,
  par = NULL,
  funcs = NULL,
  env = NULL,
  vals = FALSE,
  digits = 2,
  plot.numbers = TRUE,
  plot.cors = TRUE,
  keep.par = FALSE,
  ...
)
}
\description{
Plot taxis2
}
