% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.tracks}
\alias{plotmomo.tracks}
\title{Plot tracks}
\usage{
plotmomo.tracks(x, n = 10)
}
\arguments{
\item{x}{list of tracks}

\item{n}{Number of samples}
}
\value{
Nothing.
}
\description{
Plotting function that allows to plot tracks.
}
