% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prep.effort}
\alias{prep.effort}
\title{Prepare effort data}
\usage{
prep.effort(x, date.format = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Either a list with environmental data at different time points as
matrices or an array where the third dimension corresponds to the time
points.}

\item{date.format}{Optional; allows to specify the format of the time points
(either names of list elements or attributes for the third dimension
depending on the format of x).}

\item{verbose}{if \code{TRUE}, print information to console. Default: \code{TRUE}.}
}
\value{
A 3-dimensional array with the gridded effort data for x and
y location (first two dimnesions) and the time points as third
dimension.
}
\description{
\code{prep.effort} checks a list with effort information and
converts it into the object required by \emph{momo}.
}
