% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prep.env}
\alias{prep.env}
\title{Prepare environmental data}
\usage{
prep.env(x, date.format = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Either a list with environmental data at different time points as
matrices or an array where the third dimension corresponds to the time
points.}

\item{date.format}{Optional; allows to specify the format of the time points
(either names of list elements or attributes for the third dimension
depending on the format of x).}

\item{verbose}{if \code{TRUE}, print information to console. Default: \code{TRUE}.}
}
\value{
A 3-dimensional array with the gridded environmental data for x and
y location (first two dimnesions) and the time points as third
dimension.
}
\description{
\code{prep.env} checks a list with information about mark-resight
tags and converts it into the object required by \emph{momo}.
}
\examples{

data(skjepo)

env <- prep.env(skjepo.env)

}
