% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prep.stags}
\alias{prep.stags}
\title{Prepare mark-resight tags}
\usage{
prep.stags(x, names, origin = NULL, speed.limit = NULL, verbose = TRUE)
}
\arguments{
\item{x}{List with data frames where each data frame holds information about
the track of one mark-resight tag. At a minimum, each data frame has to
include a column for the time and x and y position of a tag.}

\item{names}{Names of columns that contain in following order: time,
location x, location y.}

\item{origin}{Optional; allows to convert a date from a numeric format to
date format by using the origin. Default: \code{NULL}.}

\item{speed.limit}{Optional; allows to apply a speed filter in km/d. All
tags that would imply a larger speed by assuming the distance between
recapture and relase location. Default: \code{NULL}.}

\item{verbose}{if \code{TRUE}, print information to console. Default: \code{TRUE}.}
}
\value{
A list with prepared tags.
}
\description{
\code{prep.stags} checks a list with information about mark-resight
tags and converts it into the object required by \emph{momo}.
}
