% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rescale.env}
\alias{rescale.env}
\title{Rescale an env field}
\usage{
rescale.env(env, zrange, i = NULL)
}
\arguments{
\item{env}{field}

\item{zrange}{range}

\item{i}{NULL}
}
\value{
Rescaled env field
}
\description{
Rescale an env field
}
