% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_MOMOmaster_set_and_run.R
\name{RunMoMo}
\alias{RunMoMo}
\title{Runs the MoMo code.}
\usage{
RunMoMo()
}
\description{
The master start button for running all of the MoMo code. \code{\link{SetOpts}} must be run first.
}
\examples{
SetOpts(
  DoA=as.Date("2013-12-31"),
  DoPR=as.Date("2008-1-1"),
  WStart=1,
  WEnd=52,
  country = "Denmark",
  source = "SSI",
  MFILE = "DoD_DoR.txt",
  HFILE = "holidays.txt",
  INPUTDIR = system.file("testdata",package="MOMO"),
  WDIR = tempdir(),
  back = 3,
  WWW = 290,
  Ysum = 2013,
  Wsum = 40,
  USEglm2 = TRUE,
  useAUTOMN = TRUE,
  datesISO = FALSE,
  plotGraphs = FALSE
)

#RunMoMo()
}
