% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_MOMOmaster_set_and_run.R
\name{SetOpts}
\alias{SetOpts}
\title{Set all the options for MoMo.}
\usage{
SetOpts(DoA = as.Date("2015-8-10"), DoPR = as.Date("2008-1-1"),
  WStart = 1, WEnd = 52, country = "Greece", source = "UoP",
  MFILE = "greece_m.dta", HFILE = "holidayfilegreece.dta",
  INPUTDIR = "./input", MDATA = NULL, HDATA = NULL,
  WDIR = "./output", back = 3, WWW = 290, Ysum = 2009, Wsum = 34,
  USEglm2 = TRUE, useAUTOMN = TRUE, datesISO = FALSE,
  plotGraphs = TRUE, removeDataAfterDoA = TRUE,
  delayVersion = "original", delayFunction = NULL,
  delayVariance = FALSE, MOMOgroups = list(`0to4` =
  "age >= 0 & age <=4", `5to14` = "age >= 5 & age <=14", `15to64` =
  "age >= 15 & age <=64", `65P` = "age >= 65 | is.na(age)", Total =
  "age >= 0 | is.na(age)"), MOMOmodels = c(`0to4` = "LINE", `5to14` =
  "LINE", `15to64` = "LINE_SIN", `65P` = "LINE_SIN", Total = "LINE_SIN"),
  verbose = TRUE)
}
\arguments{
\item{DoA}{Date of aggregation (see specifications, the only information to change weekly). Provided in ISO format, i.e. YYYY-MM-DD.}

\item{DoPR}{Date of start of a regular MOMO registration (see specifications). Provided in ISO format, i.e. YYYY-MM-DD.}

\item{WStart}{Week of cumulative excess start. e.g. "influenza season" as define by EISS will be defined as WStart = 40, WEnd= 20. e.g. summer could be defined as WStart = 26, WEnd= 40}

\item{WEnd}{Week of cumulative excess end e.g. "influenza season" as define by EISS will be defined as WStart = 40, WEnd= 20.}

\item{country}{Country name}

\item{source}{Source of data}

\item{MFILE}{Name of your mortality file, either stata format (.dta) or text file (.txt) [you can use MDATA instead]}

\item{HFILE}{Name of file containing bank Holidays (see specifications), either stata format (.dta) or text file (.txt) [you can use HDATA instead]}

\item{INPUTDIR}{Input directory (where MFILE and HFILE can be found)}

\item{MDATA}{A data.frame containing the data that would have been contained inside MFILE. MDATA takes priority over MFILE if both are specified.}

\item{HDATA}{A data.frame containing the data that would have been contained inside HFILE. HDATA takes priority over HFILE if both are specified.}

\item{WDIR}{Output directory (all output will go here)}

\item{back}{The number of weeks to remove for modeling delay = the part of the series that require delay correction (see specifications).}

\item{WWW}{Length of retrospective historical study period in weeks}

\item{Ysum}{START OF CUSUM CHART: Week for CUSUM to be set to 0}

\item{Wsum}{START OF CUSUM CHART: Week for CUSUM to be set to 0}

\item{USEglm2}{Use glm2() in order to improve convergence properties}

\item{useAUTOMN}{Keep using the column name "Automn" (instead of "Autumn") as in Stata MOMOpack}

\item{datesISO}{When saving dates in text files, use ISO format (standard in R) instead of the Stata "\%d" format}

\item{plotGraphs}{Setting this to FALSE suppressess the plotting of the various graphs (and saves time)}

\item{removeDataAfterDoA}{Set this to FALSE to include people who were registered/died after DoA}

\item{delayVersion}{Using the "original" delay code or the "2017-12" delay code}

\item{delayFunction}{This can be used instead of delayVersion. It lets you use a custom delay correction function.}

\item{delayVariance}{If this is true, it means that the prediction variance (from delay correction) is included in the z-scores+confidence interval}

\item{MOMOgroups}{Definition of the groups to be analyzed}

\item{MOMOmodels}{Names in the following vector should correspond to the MOMOgroups and the corresponding values (model to use for each group) should be one of "LINE", "SPLINE", "LINE_SIN", "SPLINE_SIN"}

\item{verbose}{Printing out information}
}
\description{
It is usual to run \code{\link{RunMoMo}} after these options have been set.
}
