% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setup.momo.data}
\alias{setup.momo.data}
\title{setup.momo.data}
\usage{
setup.momo.data(
  grid,
  env,
  ctags = NULL,
  atags = NULL,
  effort = NULL,
  knots.tax = NULL,
  knots.dif = NULL,
  const.dif = TRUE,
  trange = NULL
)
}
\description{
setup.momo.data
}
