% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.alpha}
\alias{sim.alpha}
\title{Simulate alpha}
\usage{
sim.alpha(ll, ul, n = 3, digits = 2)
}
\arguments{
\item{ll}{lower limit of \link{runif} to simulate parameters.}

\item{ul}{upper limit of \link{runif} to simulate parameters.}

\item{n}{number of alpha parameters. Default: \code{3}.}

\item{digits}{number of significant digits for the parameters. Default: \code{2}.}
}
\value{
A vector with simulate alpha values.
}
\description{
Simulate alpha parameter specifying taxis component.
}
