% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.atags}
\alias{sim.atags}
\title{Simulate archival tags}
\usage{
sim.atags(
  grid,
  par = NULL,
  env = NULL,
  n = 30,
  effort = NULL,
  trange = NULL,
  dt = NULL,
  trange.rel = NULL,
  xrange.rel = NULL,
  yrange.rel = NULL,
  trange.rec = NULL,
  by = 0.1,
  knots.tax = NULL,
  knots.dif = NULL,
  funcs = NULL
)
}
\description{
Simulate archival tags
}
