% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.ctags}
\alias{sim.ctags}
\title{Simulate conventional tags}
\usage{
sim.ctags(
  grid,
  par = NULL,
  env = NULL,
  n = 100,
  effort = NULL,
  trange = NULL,
  trange.rel = NULL,
  xrange.rel = NULL,
  yrange.rel = NULL,
  trange.rec = NULL,
  by = 0.01,
  knots.tax = NULL,
  knots.dif = NULL,
  funcs = NULL
)
}
\description{
Simulate conventional tags
}
