% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.env}
\alias{sim.env}
\title{Simulate environmental fields}
\usage{
sim.env(
  grid,
  nt = 9,
  rho_t = 0.85,
  sd = 2,
  h = 0.2,
  nu = 2,
  rho_s = 0.8,
  delta = 0.1,
  zrange = c(20, 28),
  matern = TRUE,
  diagonal = FALSE
)
}
\description{
Simulate environmental fields
}
