% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.momo}
\alias{sim.momo}
\title{Simulate all data needed for fit.momo}
\usage{
sim.momo(
  fit = NULL,
  xrange = c(0, 1),
  yrange = c(0, 1),
  dxdy = c(0.1, 0.1),
  select = FALSE,
  plot.land = FALSE,
  keep.par = FALSE,
  nt = 1,
  rho_t = 0.85,
  sd = 2,
  h = 0.2,
  nu = 2,
  rho_s = 0.8,
  delta = 0.1,
  zrange = c(20, 28),
  matern = TRUE,
  diagonal = FALSE,
  par = NULL,
  n.alpha = 3,
  const.dif = TRUE,
  knots.tax = NULL,
  knots.dif = NULL,
  correct.peclet = FALSE,
  use.ctags = TRUE,
  n.ctags = 300,
  use.atags = TRUE,
  n.atags = 30,
  trange.rel = NULL,
  xrange.rel = c(0.4, 0.6),
  yrange.rel = c(0.4, 0.6),
  verbose = TRUE
)
}
\arguments{
\item{fit}{optional. momo fit}
}
\description{
Simulate all data needed for fit.momo
}
