% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momo.R
\name{TECH1}
\alias{TECH1}
\title{Present Mplus results with matching numbers on TECH1 outputs from Mplus}
\usage{
TECH1(mplusoutput.file = "example_d3.out")
}
\arguments{
\item{mplusoutput.file}{Mplus output file (with TECH 1 on the OUTPUT: command in Mplus input file).}
}
\value{
outputs on screen
}
\description{
Present Mplus estimated parameters with matching number on TECH1 outputs (for inputs in momo_mplus)
}
\examples{

## -- Example -- ##

# Mplus output file "example_d3.out" & output TECH1 object is Tech1

Tech1 <- TECH1("example_d3.out")

}
