% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create.grid}
\alias{create.grid}
\title{Create a grid}
\usage{
create.grid(
  xrange = c(0, 1),
  yrange = c(0, 1),
  dxdy = c(0.1, 0.1),
  select = FALSE,
  plot.land = FALSE,
  keep.par = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{xrange}{Range x}

\item{yrange}{Range y}

\item{dxdy}{dxdy}

\item{plot.land}{Optional plot land (only when select.cells = TRUE)}

\item{verbose}{Print messages?}

\item{select.cells}{Select cells}
}
\description{
Create a grid
}
