% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{def.conf}
\alias{def.conf}
\title{Generate default configuration list}
\usage{
def.conf(dat)
}
\arguments{
\item{dat}{data frame with input data as produced by the function
\link{check.momo.data}.}
}
\value{
A list with configurations.
}
\description{
\code{def.conf} generates a list with default configurations
required by the function \link{fit.momo}.
}
\examples{
data(skjepo)
conf <- def.conf(skjepo$dat)

}
