% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{def.map}
\alias{def.map}
\title{def.map}
\usage{
def.map(dat, conf, par)
}
\arguments{
\item{dat}{Data list with required information}

\item{conf}{Configuration list}

\item{par}{parameter list}
}
\description{
def.map
}
