% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.momo}
\alias{fit.momo}
\title{Fit the movement model}
\usage{
fit.momo(
  dat,
  conf = NULL,
  par = NULL,
  map = NULL,
  newtonsteps = 3,
  rm.unidentified = FALSE,
  run = TRUE,
  lower = get.lower.bounds(par),
  upper = get.upper.bounds(par),
  sim.condRE = TRUE,
  ignore.parm.uncertainty = FALSE,
  rel.tol = 1e-10,
  verbose = TRUE,
  do.sdreport = TRUE,
  do.report = TRUE,
  use.expm = NULL,
  ...
)
}
\arguments{
\item{dat}{list of dat objects}

\item{conf}{list of initial model parameters}

\item{par}{list of initial model parameters}
}
\value{
list of result of minimization conducted via nlminb
}
\description{
Fit the movement model
}
\details{
...
}
\examples{
3 + 3
}
