% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.dim}
\alias{get.dim}
\title{Get dimensions from tagging data}
\usage{
get.dim(ctags = NULL, atags = NULL)
}
\arguments{
\item{ctags}{Conventional tags}

\item{atags}{Archival tags}
}
\description{
Get dimensions from tagging data
}
