% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{get.lower.bounds}
\alias{get.lower.bounds}
\title{Lower bounds}
\usage{
get.lower.bounds(par)
}
\arguments{
\item{par}{parameter list with initial values as produced by the function
\link{def.par}.}
}
\value{
List with lower bounds for parameters.
}
\description{
Get lower bounds for parameters.
}
