% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.neighbours}
\alias{get.neighbours}
\title{Get neighbours for each cell of a grid}
\usage{
get.neighbours(celltable, diagonal = FALSE)
}
\arguments{
\item{celltable}{data frame with numbered cells as returned by the function
\link{create.grid}.}

\item{diagonal}{logical; if \code{TRUE}, diagonal neighbours are included.
Default: \code{FALSE}.}
}
\value{
A matix with each cell and corresponding neighbouring cells.
}
\description{
\code{get.neighbours} allows to get a
}
\examples{
grid <- create.grid()

neighbours <- get.neighbours(grid$celltable)

}
