% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.precision.matrix}
\alias{get.precision.matrix}
\title{Get precision matrix}
\usage{
get.precision.matrix(grid, h, nu, rho, delta, matern = TRUE, diagonal = FALSE)
}
\arguments{
\item{grid}{Grid}

\item{delta}{Parameter}
}
\description{
Get precision matrix
}
