% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get.upper.bounds}
\alias{get.upper.bounds}
\title{Upper bounds}
\usage{
get.upper.bounds(parameters)
}
\arguments{
\item{parameters}{initial values for the model in a format similar to what
is returned from the defpar function}
}
\value{
a named list
}
\description{
Upper bounds
}
