% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{model.version.info}
\alias{model.version.info}
\title{Description of model}
\usage{
model.version.info(fit, ...)
}
\arguments{
\item{fit}{A fitted momo list of class \code{momo.fit} as returned by the
function \link{fit.momo}.}

\item{...}{Additional parameters to be passed to \link{writeLines}.}
}
\description{
Writes a string to install the version of the package which was
used to run the model.
}
