% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccimm.R
\name{momo_mplus_fun}
\alias{momo_mplus_fun}
\title{Monte Carlo Simulation for Confidence Intervals of Defined Function (mplus)}
\usage{
momo_mplus_fun(
  mplus_output_file = "mplus_output.out",
  results_file = "results.txt",
  a1 = "NA",
  a2 = "NA",
  a3 = "NA",
  a4 = "NA",
  z1 = "NA",
  z2 = "NA",
  z3 = "NA",
  z4 = "NA",
  w1 = "NA",
  w2 = "NA",
  w3 = "NA",
  w4 = "NA",
  zw1 = "NA",
  zw2 = "NA",
  zw3 = "NA",
  zw4 = "NA",
  Sfunction = "NULL",
  R = 5
)
}
\arguments{
\item{mplus_output_file}{Mplus output (.out) file (output from Mplus).}

\item{results_file}{Mplus a text file (.txt) that saves the Mplus results (RESULTS IS "filename.txt" in Mplus SAVEDATA:).}

\item{a1}{location of parameter a1 in Mplus Tech1 outputs.}

\item{a2}{location of parameter a2 in Mplus Tech1 outputs.}

\item{a3}{location of parameter a3 in Mplus Tech1 outputs.}

\item{a4}{location of parameter a4 in Mplus Tech1 outputs.}

\item{z1}{location of parameter z1 in Mplus Tech1 outputs.}

\item{z2}{location of parameter z2 in Mplus Tech1 outputs.}

\item{z3}{location of parameter z3 in Mplus Tech1 outputs.}

\item{z4}{location of parameter z4 in Mplus Tech1 outputs.}

\item{w1}{location of parameter w1 in Mplus Tech1 outputs.}

\item{w2}{location of parameter w2 in Mplus Tech1 outputs.}

\item{w3}{location of parameter w3 in Mplus Tech1 outputs.}

\item{w4}{location of parameter w4 in Mplus Tech1 outputs.}

\item{zw1}{location of parameter zw1 in Mplus Tech1 outputs.}

\item{zw2}{location of parameter zw2 in Mplus Tech1 outputs.}

\item{zw3}{location of parameter zw3 in Mplus Tech1 outputs.}

\item{zw4}{location of parameter zw4 in Mplus Tech1 outputs.}

\item{Sfunction}{function of estimated parameters from Mplus outputs.}

\item{R}{number of Monte Carlo simulation samples (in millions). For example, R=5 (default) generates 5,000,000 simulated samples.}
}
\value{
confidence intervals of defined function
}
\description{
Generate confidence intervals of a defined function from Mplus results using Monte Carlo simulation.
Location of estimated parameters can be found in momo::TECH1().
}
\examples{

## -- Example -- ##

# mplus_output_file is "model cc4.out", results_file is "Model_CC4.txt" & moderator Z is "AUTO"
# output momo object is mcObject

mcObject <- momo_mplus_fun("model cc4.out", "Model_CC4.txt",
            a1 = "60", a2 = "65", Sfunction="a1*a2")

}
