% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{nll}
\alias{nll}
\title{Negative log likelihood function}
\usage{
nll(par, dat)
}
\arguments{
\item{par}{a list of initial parameters}

\item{dat}{the data set}
}
\value{
a scalar function value
}
\description{
Negative log likelihood function
}
\details{
exported, but mainly intended for internal use
}
