% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.land}
\alias{plot.land}
\title{Plot land masses}
\usage{
\method{plot}{land}(xlim, ylim)
}
\arguments{
\item{xlim}{x limits}

\item{ylim}{y limits}
}
\value{
Nothing.
}
\description{
Plot land masses.
}
