% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.atags}
\alias{plotmomo.atags}
\title{Plot momo atags}
\usage{
plotmomo.atags(
  x,
  main = "Archival tags",
  keep.par = FALSE,
  plot.land = FALSE,
  xlim = NULL,
  ylim = NULL,
  xlab = "x",
  ylab = "y",
  leg.pos = "topright",
  ...
)
}
\description{
Plot momo atags
}
