% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.compare.one}
\alias{plotmomo.compare.one}
\title{Plot to compare single quantity}
\usage{
plotmomo.compare.one(
  fit,
  ...,
  quantity = c("pref", "taxis", "dif", "par"),
  keep.gpar = FALSE,
  asp = 2,
  col = momo.cols(10),
  lty = 1:10,
  plot.legend = 1
)
}
\arguments{
\item{fit}{a list of class \code{momo.fit} as returned by \link{fit.momo} or a list of
such (named) fitted objects.}

\item{...}{optional; additional momo fits.}

\item{quantity}{quantities that can be compared. The following options are
currently implemented: \code{"taxis"} for the taxis, \code{"dif"} for the
diffusion, \code{"par"} for the parameters.}

\item{asp}{positive number deining the target aspect ratio (columns / rows)
of the plot arrangement.}

\item{col}{colors. By default, \link{momo.cols} are used: \code{momo.cols(10)}.}

\item{lty}{vector with line types. Default: 1:10.}

\item{plot.legend}{logical or integer indicating which legend should be
plotted. Default: \code{1}.}
}
\value{
Nothing.
}
\description{
Plotting function that allows to compare multiple fits or
fitted and simulated data.
}
