% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.compare}
\alias{plotmomo.compare}
\title{Compare plot}
\usage{
plotmomo.compare(
  fit,
  ...,
  quantity = c("pref", "taxis", "dif", "par"),
  keep.gpar = FALSE,
  col = momo.cols(10),
  asp = 2,
  plot.legend = 1
)
}
\arguments{
\item{fit}{a list of class \code{momo.fit} as returned by \link{fit.momo} or a list of
such (named) fitted objects.}

\item{...}{optional, additional momo fits.}

\item{quantity}{quantities that can be compared. The following options are
currently implemented: \code{"taxis"} for the taxis, \code{"dif"} for
the diffusion, \code{"par"} for the parameters.}

\item{keep.gpar}{logical; If \code{TRUE}, do not overwrite the graphical
parameters. Default: \code{FALSE}.}

\item{asp}{positive number deining the target aspect ratio (columns / rows)
of the plot arrangement.}

\item{plot.legend}{logical or integer indicating which legend should be
plotted. Default: \code{1}.}
}
\value{
Nothing.
}
\description{
Plotting function that allows to compare multiple fits or
fitted and simulated data.
}
