% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.dif}
\alias{plotmomo.dif}
\title{Plot diffusion}
\usage{
plotmomo.dif(
  x,
  cor = 20,
  main = "Diffusion",
  col = "black",
  alpha = 0.5,
  lwd = 1,
  keep.par = FALSE,
  add = FALSE,
  ...
)
}
\description{
Plot diffusion
}
