% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.effort}
\alias{plotmomo.effort}
\title{Plot momo effort}
\usage{
plotmomo.effort(
  effort,
  main = "Effort fields",
  labels = TRUE,
  keep.par = FALSE,
  plot.land = FALSE,
  xlab = "x",
  ylab = "y",
  ...
)
}
\description{
Plot momo effort
}
