% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.grid}
\alias{plotmomo.grid}
\title{Plot momo grid}
\usage{
plotmomo.grid(
  grid,
  main = "Grid",
  labels = TRUE,
  plot.land = FALSE,
  keep.par = FALSE,
  ...
)
}
\description{
Plot momo grid
}
