% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmomo.taxis}
\alias{plotmomo.taxis}
\title{Plot taxis}
\usage{
plotmomo.taxis(
  x,
  cor = 1,
  par = NULL,
  funcs = NULL,
  env = NULL,
  alpha = 0.5,
  main = "Taxis",
  keep.par = FALSE,
  ...
)
}
\description{
Plot taxis
}
