% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prep.env}
\alias{prep.env}
\title{prep.env}
\usage{
prep.env(x, names, origin = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Data frame with required information}

\item{names}{Names of columns that contain in following order: release time (t0), recapture time (t1), release location x (x0), recapture location x (x1), release location y (y0), recapture location y (y1)}

\item{origin}{Optional, if time column not provided as decimal date, origin can be used to convert date columns to dates}

\item{verbose}{Print messages?}
}
\description{
prep.env
}
