% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.env}
\alias{sim.env}
\title{Simulate environmental fields}
\usage{
sim.env(
  grid,
  nt = 9,
  rho_t = 0.85,
  sd = 2,
  h = 0.2,
  nu = 2,
  rho_s = NULL,
  delta = 0.1,
  zrange = c(20, 28),
  matern = TRUE,
  diagonal = FALSE,
  sim.buffer = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{grid}{a grid.}

\item{nt}{number of environmental fields. Default: \code{1}.}

\item{rho_t}{autocorrelation coefficient of temporal autocorrelation for
environmental fields. Default: \code{0.85}.}

\item{sd}{standard deviation for simulation of environmental fields.
Default: \code{2}.}

\item{h}{parameter of the matern covariance structure. Default: \code{0.2}.}

\item{nu}{parameter of the matern covariance structure. Default: \code{2}.}

\item{rho_s}{parameter of the matern covariance structure. Default: \code{0.8}.}

\item{delta}{parameter of the matern covariance structure. Default: \code{0.1}.}

\item{zrange}{range of the environmental covariate. Default: \code{c(20, 28)}.}

\item{matern}{logical; if \code{TRUE} (default), matern covariance structure is
used for simulation of environmental fields.}

\item{diagonal}{logical; if \code{TRUE}, diagonal neighbours are considered in
neighbouring structure. Default: \code{FALSE}.}

\item{sim.buffer}{logical; if \code{TRUE} extends the provided grid so that the
2D interpolation function does not return \code{NaN} close to the boundary.}

\item{verbose}{If \code{TRUE}, print information to console. Default: \code{TRUE}.}
}
\value{
A list with simulated environmental fields.
}
\description{
\code{sim.env} allows to simulate environmental fields.
}
\examples{
env <- sim.env()

}
